Refactor validation back into telemetry api methods.
